+++ CHANGELOG.md
@@ -17,7 +17,7 @@
 - refactor
 
 
-## v1.15-SNAPSHOT
+## v1.15-SNAPSHOT / 2024-11-28
 
 - changed:home/docstring --- take the last docstring found.